channel-provider: moved JsonRpc types, added tests
