+++ packages/embedded-wallet/src/message-dispatchers/allocate/allocate.test.ts
@@ -1,5 +1,5 @@
-import {JsonRpcResponse} from '../../types';
-import {allocate} from './allocate';
+import {JsonRpcErrorResponse, JsonRpcResponse} from '@statechannels/channel-provider';
+import {allocate, rejectAllocation} from './allocate';
 
 describe('MessageDispatchers - Allocate', () => {
   it('should dispatch a JSONRPC call with the `allocate` method', async done => {
@@ -16,5 +16,24 @@ describe('MessageDispatchers - Allocate', () => {
       done();
     };
     allocate(123, {allocated: true});
+  });
+
+  it('should dispatch a JSONRPC call with the `rejectAllocation` method', async done => {
+    window.onmessage = (event: MessageEvent) => {
+      const response = event.data as JsonRpcErrorResponse;
+      if (!response.jsonrpc) {
+        fail('Expected a JSONRPC response');
+      } else {
+        expect(response.jsonrpc).toEqual('2.0');
+        expect('result' in response).toEqual(false);
+        expect(response.error).toEqual({
+          code: -32100,
+          message: 'User has rejected budget allocation'
+        });
+        expect(response.id).toEqual(123);
+      }
+      done();
+    };
+    rejectAllocation(123);
   });
 });
