--- packages/embedded-wallet/src/message-dispatchers/error-codes.ts
@@ -1,3 +1,0 @@
-export enum JsonRpcErrorCodes {
-  BudgetAllocationRejected = -32100
-}
